+++ client/pages/Warehouse/shared/StepperInput.tsx
@@ -88,6 +88,7 @@ export const StepperInput = forwardRef<HTMLInputElement, StepperInputProps>(
             isDisabled={disabled}
             className="absolute left-2 top-1/2 -translate-y-1/2 h-14 w-10 min-w-6 z-10 bg-gray-100"
             onPress={onDecrement}
+            tabIndex={-1}
             aria-label="Зменшити"
           >
             <DynamicIcon name="minus" className="w-6 h-6" />
@@ -98,6 +99,7 @@ export const StepperInput = forwardRef<HTMLInputElement, StepperInputProps>(
             isDisabled={disabled || (max !== undefined && value >= max)}
             className="absolute right-2 top-1/2 -translate-y-1/2 h-14 w-10 min-w-6 z-10 bg-gray-100"
             onPress={onIncrement}
+            tabIndex={-1}
             aria-label="Збільшити"
           >
             <DynamicIcon name="plus" className="w-6 h-6" />